requirements.txt
flask
PyPDF2
Pillow
pandas
openpyxl
python-docx
werkzeug